% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanColNames.R
\name{cleanColNames}
\alias{cleanColNames}
\title{Tidy up column names}
\usage{
cleanColNames(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
a (tidied) data frame
}
\description{
Removes redundant punctuation and whitespace from data frame
}
