% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSubstructure.R
\name{createSubstructure}
\alias{createSubstructure}
\title{Create typical project subfolders}
\usage{
createSubstructure(location = getwd(), folders = c("analysis", "src",
  "data", "output", "doc", "ext"))
}
\arguments{
\item{location}{where to create the subfolders. Defaults to working directory}

\item{folders}{vector of strings specifying the names of folders to create}
}
\value{
logical vector confirming whether folder was created or not
}
\description{
Create typical project subfolders
}
