% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeStamp.R
\name{dateTimeStamp}
\alias{dateTimeStamp}
\title{A simple function to prefix a character vector with \code{ymd_hms}}
\usage{
dateTimeStamp(character)
}
\arguments{
\item{character}{a character vector}
}
\value{
a character vector, prefixed with a timestamp
}
\description{
A simple function to prefix a character vector with \code{ymd_hms}
}
