% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagesToGIF.R
\name{imagesToGIF}
\alias{imagesToGIF}
\title{Convert a directory of images to .gif using \code{magick}}
\usage{
imagesToGIF(image_directory = NULL, output_filename = NULL,
  scale = NULL, filetype = "png", ...)
}
\arguments{
\item{image_directory}{directory including images whose filenames, when sorted, are in order. Opens dialog if left \code{NULL}.}

\item{output_filename}{name of output file. Opens dialog if left \code{NULL}.}

\item{scale}{specify size (in pixels) of longest dimension of image}

\item{filetype}{filetype of image file. Defaults to \code{png}}

\item{...}{additional arguments to be passed to \code{image_animate}}
}
\value{
returns nothing in the environment - but creates gif of images externally
}
\description{
Convert a directory of images to .gif using \code{magick}
}
