% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAllSheets.R
\name{readAllSheets}
\alias{readAllSheets}
\title{Read all sheets from an xlsx or xls file into R as a list}
\usage{
readAllSheets(filename, ...)
}
\arguments{
\item{filename}{the path of an \code{xlsx} or \code{xls} file.}

\item{...}{additional arguments to be passed to \link[readxl]{read_xlsx}.}
}
\value{
a list of data frames: one for each sheet in the source file
}
\description{
Simple function which reads all sheets from an \code{xlsx} or \code{xls} file into a list
}
